Transformations-Projections-Rasterization/requirements.txt
numpy
tqdm
matplotlib.pyplot
imageio
opencv-python
